Test is passing to check the functionality of textarea's onChange method
Here we add a constructor with state of 'value' to our class component WhistleForm.
Then, add the initial value of an empty string to our textarea tag.
Then, add an onChange method prop to our textarea tag, which invokes the
handleChange method of our component, which sets the state of our value
key to be updated with whatever text is inputted into the textarea by
the user.