Test is passing for functionality of submission of a new whistle
Here, I add state to the App component of an empty 'whistles' array.
Then, pass down a prop 'addWhistles' to the WhistleForm component. Next,
in WhistleForm add a method handleSubmit which invokes the addWhistle
method passed down as props. The argument of addWhistle is the state of
our WhistleForm representing an object. The state now includes a
userSelected key which is updated by the handleUserChange method, which
from a newly created select menu tag with options of our users, lets the
user choose their name before constructing a whistle. Pressing the
submit button now sucessfully adds the new whistle to the whistles array
in App to later be rendered to the timeline